feat: add debounce helper, and chore: move validation helper to good spot
